requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2